added sorta working demo
character controller needs work, it's still buggy
