+++ Project Ronin/Assets/Scripts/Ability System/AttributeInitializer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d53f25bdeba74a14ab0aba65f65ecddb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
