+++ Project Ronin/Assets/Scripts/Character Controlling/DummyCharacterFSM.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c0f84fd91ad19204a9a8a53620c94915
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
