+++ Project Ronin/Assets/Scripts/Character Controlling/DummyInput.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b7773762e3afc27438f1299be370c4ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
