+++ Project Ronin/Assets/Scripts/Johnny's Demo/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8bf14dd228f73784d9bc10628fd5036b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
