+++ Project Ronin/Assets/Scripts/Johnny's Demo/Animations/attack_clip.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2ca59a6f7c7241448b3719961f6dc586
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
