+++ Project Ronin/Assets/Scripts/Johnny's Demo/Animations/dummy.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7f1a16b813ce24a449281dbca70a07c5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
