+++ Project Ronin/Assets/Scripts/Johnny's Demo/Animations/idle_clip.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1d36312dbda738042942bf0d78259614
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
