+++ Project Ronin/Assets/Scripts/Johnny's Demo/Animations/short_attack.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 99f3cfad7dd02f34cbd832266a650a7c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
