+++ Project Ronin/Assets/Scripts/Johnny's Demo/Animations/walk_inplace.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 99b071de0bf870d42a71bbc73c2bf751
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
