+++ Project Ronin/Assets/Scripts/Johnny's Demo/Animations/walking_clip.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 68e6580b8a8775b408b618b942536f60
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
