+++ Project Ronin/Assets/Scripts/Johnny's Demo/DummyEnemy.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d53f25bdeba74a14ab0aba65f65ecddb, type: 3}
+  m_Name: DummyEnemy
+  m_EditorClassIdentifier: 
+  floatInitials:
+  - name: hp
+    value: 100
+  tagInitials: []
