+++ Project Ronin/Assets/Scripts/Johnny's Demo/DummyEnemy.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8f51e4c5b54dde5429c226f22a6c6b4d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
