+++ Project Ronin/Assets/Scripts/Johnny's Demo/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7d063ef18a4035249879523239eb1fbe
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
