+++ Project Ronin/Assets/Scripts/Johnny's Demo/Scripts/Dummy Attack.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9ae0a38e05b1fcf4c885bcfa04aa276b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
