+++ Project Ronin/Assets/Scripts/Johnny's Demo/Scripts/DummyAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dbc56866cc3408546a8bf6fa0ca54187
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
